Update book.js with regex fix and extra data to elements (#3)
* Add data to words

Adds verse, chapter, and word to the span

now printing the anchor and focus to the console

* Regex fix

had brackets around a group rather than parentheses, which was cutting
off the last number of a strong's number. Had been hacking around that,
but this is the proper way to do it.